Maximum of 3 Float Numbers completed

+++ src/com/generics/MaxNumber.java
@@ -1,16 +1,16 @@
 package com.generics;
 
 public class MaxNumber {
-		private Integer a,b,c;
+		private Float a,b,c;
 
-		public MaxNumber(Integer a, Integer b, Integer c) {
+		public MaxNumber(Float a, Float b, Float c) {
 			this.a = a;
 			this.b = b;
 			this.c = c;
 		}
 		
 		public void findMax() {
-			Integer max = a;
+			Float max = a;
 			if(b.compareTo(max)>0) {
 				max = b;
 			}
